requirements-build.txt
# Song Editor 3 - Build Requirements
# Optimized for PyInstaller packaging

# Core dependencies
PySide6>=6.5.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
librosa>=0.10.0

# Audio processing
soundfile>=0.12.0
noisereduce>=3.0.0
pyloudnorm>=0.4.1
demucs>=4.0.0

# Transcription models
openai-whisper>=20230314
faster-whisper>=1.0.0
whisperx>=3.0.0

# Analysis
chord-extractor>=0.1.0
crepe>=0.0.16
basic-pitch>=0.4.0
pretty-midi>=0.2.10
mido>=1.3.0

# Additional utilities
pathlib2>=2.3.0
typing-extensions>=4.5.0
packaging>=23.0

# Build tools
pyinstaller>=6.0.0

# Optional MLX Whisper (Apple Silicon)
# mlx-whisper>=0.1.0  # Uncomment if building for Apple Silicon
